Outfactor submission logic to useFormState
