+++ frontend/composables/useFormDataToMutationInput.ts
@@ -1,0 +1,41 @@
+import type { UserFormInput } from "~/generated/gql/graphql";
+
+/**
+ * Recursively collects all questions from a step and its substeps, and puts them in a flat array.
+ */
+function getAllQuestions(
+    step: StepWithValues | SubstepWithValues,
+): QuestionWithValue[] {
+    const ownQuestions = step.questions;
+
+    const subStepQuestions =
+        "substeps" in step
+            ? (step.substeps?.flatMap(getAllQuestions) ?? [])
+            : [];
+
+    return [...ownQuestions, ...subStepQuestions];
+}
+
+/**
+ * Utility function to transform our form into the expected input for our mutation.
+ */
+function useFormDataToMutationInput(
+    formData: FormWithValues,
+    submissionId: string | null,
+    formId: string | null,
+): UserFormInput {
+    const questions = formData.steps.flatMap(getAllQuestions);
+
+    return {
+        submissionId: submissionId,
+        formConfigId: formId,
+        responses: questions.map((question) => ({
+            answer: JSON.stringify({ value: question.value }),
+            id: question.responseId,
+            questionId: question.questionId,
+            repeatIndex: question.repeatIndex,
+        })),
+    };
+}
+
+export default useFormDataToMutationInput;
